requirements.txt
numpy
matplotlib
scipy
prody
pandas
biopython
